requirements.txt
Flask
gunicorn
sklearn
pandas
numpy
typing
PyYAML

